Moved CourseSelectColumn into SchedulingPage folder
