+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/CollapsedCourseCard/CollapsedCourseCard.css.d.ts
