+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/ExpandedCourseCard.tsx
@@ -10,9 +10,9 @@ import { useSelector, useDispatch } from 'react-redux';
 import * as styles from './ExpandedCourseCard.css';
 import SectionSelect from './SectionSelect/SectionSelect';
 import BasicSelect from './BasicSelect/BasicSelect';
-import { CustomizationLevel, CourseCardOptions } from '../../../../types/CourseCardOptions';
-import { RootState } from '../../../../redux/reducer';
-import { updateCourseCard, removeCourseCard } from '../../../../redux/actions/courseCards';
+import { CustomizationLevel, CourseCardOptions } from '../../../../../types/CourseCardOptions';
+import { RootState } from '../../../../../redux/reducer';
+import { updateCourseCard, removeCourseCard } from '../../../../../redux/actions/courseCards';
 
 interface ExpandedCourseCardProps {
   onCollapse: Function;
